Battle System Quarterly Done
The battle system chooses which player is the Pharaoh and which is the Slave randomly, as well as gives each player their card, the positions of the cards in the hands are shuffled to make it look more like they picked the cards up. Begin phase and Enemy Play phase completed in short.